Read parameters to badger items from config file
This commit is the first part to adress issue #6. It implements reading of
arguments from a configuration file using the standard library's configparser
module.

This constitutes a breaking change, since the command-line argument <item> is
now interpreted as a callable that returns a processable item, when it is not a
module. This change allows exposing factory functions instead of instances,
which can then accept parameters parsed from configuration files.